# Allow to change autocommit attribute