@mpenkov 2023-03-30
Sorry for the late reply. Looks good, but is it possible to add tests?